==> sphingolipids_targeted/1_eda.Rmd <==
---
title: "PEGASuS Exploratory Data Analyses"
author: "Mengna Huang"
date: "2/12/2020"
output: 
  html_document: 
    toc: true
    toc_float: 
      collapsed: false
      smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}

## Packages

pkg <- c("tidyverse", "data.table", "here", "readxl", "e1071", "ggrepel", "GGally")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}


## Paths

dat_dir <- "data/"
fig_dir <- "figures/eda/"


## Filenames

rawdat_fname <- here(dat_dir, "2020-02-12_PEGASuS_SLs.xlsx")
sample_sheet_name <- "Samples"
qcinjs_sheet_name <- "QCinjs"
qcexts_sheet_name <- "QCexts"
blexts_sheet_name <- "BLexts"

manifest_fname <- here("../prelim_analysis/data/0662B_1140_Serum_Ship Out_Manifest_with Sample IDs.xlsx")

pheno_asthma_fname <- here("../prelim_analysis/data/SevereAsthmaResults.csv")
pheno_control_fname <- here("../prelim_analysis/data/ControlGroupResults.csv")


## Functions

source(here("scripts/", "fnc_table1.R"))

```

# 1. Read in data & check IDs

```{r 1.read_data}

## Sample data

samp_dat <- read_xlsx(rawdat_fname, sheet = sample_sheet_name, skip = 2) %>% as.data.table()
setnames(samp_dat, gsub(" ", "_", colnames(samp_dat)))
setnames(samp_dat, "3_KS", "X3_KS")

sph_list <- colnames(samp_dat)[(which(colnames(samp_dat) == "Comments_sample")+1):ncol(samp_dat)]
length(sph_list)

mets_miss <- samp_dat[, lapply(.SD, function(x) mean(is.na(x))), .SDcols = sph_list] %>% as.data.frame()
rownames(mets_miss) <- "pct_na_met"
mets_miss <- t(mets_miss) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(mets_miss, "rowname", "metabolite")

mets_skew <- samp_dat[, lapply(.SD, function(x) e1071::skewness(x, na.rm = T)), .SDcols = sph_list] %>% as.data.frame()
rownames(mets_skew) <- "skewness_met"
mets_skew <- t(mets_skew) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(mets_skew, "rowname", "metabolite")

mets_info <- merge(mets_miss, mets_skew, by = "metabolite")

ggplot(mets_info) + 
        geom_histogram(aes(pct_na_met*100)) + 
        labs(title = "Missingness in metabolites", x = "% missing") + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "metabolite_missingness.png"), width = 9, height = 6)

mets_info[, .N, skewness_met > 2]
ggplot(mets_info) + 
        geom_point(aes(x = rank(skewness_met, ties.method = "first"), y = skewness_met)) + 
        labs(title = "Skewness of metabolites", x = "Skewness") + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "metabolite_skewness.png"), width = 9, height = 6)


## Mean and SD

mets_mean <- samp_dat[, lapply(.SD, function(x) mean(x, na.rm = T)), .SDcols = sph_list] %>% as.data.frame()
rownames(mets_mean) <- "mean_met"
mets_mean <- t(mets_mean) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(mets_mean, "rowname", "metabolite")

mets_info <- merge(mets_info, mets_mean, by = "metabolite")

mets_sd <- samp_dat[, lapply(.SD, function(x) sd(x, na.rm = T)), .SDcols = sph_list] %>% as.data.frame()
rownames(mets_sd) <- "sd_met"
mets_sd <- t(mets_sd) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(mets_sd, "rowname", "metabolite")

mets_info <- merge(mets_info, mets_sd, by = "metabolite")


## Break into a few lists

cer_list <- grep("^Cer", sph_list, value = T) %>% sort()
sm_list <- grep("^SM", sph_list, value = T) %>% sort()
ceroth_list <- grep("^DhCer|^HexCer|^LacCer", sph_list, value = T) %>% sort()
sphoth_list <- sph_list[!(sph_list %in% c(cer_list, sm_list, ceroth_list))]

samp_dat[, asthma_sev := case_when(Group == "Asthmatic" ~ 1,
                                   Group == "Control" ~ 0)] # Add a numeric version of asthma status


## QC data

qcinjs_dat <- read_xlsx(rawdat_fname, sheet = qcinjs_sheet_name) %>% as.data.table()
setnames(qcinjs_dat, gsub(" ", "_", colnames(qcinjs_dat)))
setnames(qcinjs_dat, "3_KS", "X3_KS")
qcinjs_dat <- qcinjs_dat[!is.na(Filename)]
sph_list %in% colnames(qcinjs_dat) %>% table()

qcexts_dat <- read_xlsx(rawdat_fname, sheet = qcexts_sheet_name) %>% as.data.table()
setnames(qcexts_dat, gsub(" ", "_", colnames(qcexts_dat)))
setnames(qcexts_dat, "3_KS", "X3_KS")
qcexts_dat <- qcexts_dat[!is.na(Filename)]
sph_list %in% colnames(qcexts_dat) %>% table()


## Blank data

blexts_dat <- read_xlsx(rawdat_fname, sheet = blexts_sheet_name) %>% as.data.table()
setnames(blexts_dat, gsub(" ", "_", colnames(blexts_dat)))
setnames(blexts_dat, "3_KS", "X3_KS")
sph_list %in% colnames(blexts_dat) %>% table()


## Check consistency with sample manifest

manifest <- read_xlsx(manifest_fname, col_names = F)
colnames(manifest) <- c("LabID", "InvCode", "SubjectID", "SampleType", "Volume", "Box", "ContainerLabID", "Slot", "SampleIDBiobank")
manifest <- manifest[-c(seq(1, by = 102, length.out = 12), seq(2, by = 102, length.out = 12)), ] %>% as.data.table()
manifest[, SubjectID := as.integer(SubjectID)]
manifest[, SubjectID] %in% samp_dat[, SubjectID] %>% table()
manifest[, SubjectID] %in% samp_dat[, Match_ID] %>% table()


## Check consistency with pheno data (sent by Jessica for her grant in June)

pheno_asthma <- fread(pheno_asthma_fname)
pheno_asthma[, V1 := NULL]

pheno_control <- fread(pheno_control_fname)
pheno_control[, V1 := NULL]

identical(colnames(pheno_asthma), colnames(pheno_control))
pheno <- rbind(pheno_asthma, pheno_control)

setnames(pheno, tolower(colnames(pheno)))
setnames(pheno, "biobank.subject.id", "SubjectID")

pheno[, SubjectID] %in% samp_dat[, SubjectID] %>% table()
pheno[, SubjectID] %in% manifest[, SubjectID] %>% table()

pheno[, SubjectID][!(pheno[, SubjectID] %in% samp_dat[, SubjectID])]
pheno[, SubjectID][!(pheno[, SubjectID] %in% manifest[, SubjectID])]
pheno[SubjectID == 10037071, .(SubjectID, gender, age, race, bmi)]

```

# 2. Check QC sample CVs

```{r 2.cv}

qcinjs_dat[, (sph_list) := lapply(.SD, as.numeric), .SDcols = sph_list]
cv_qcinjs <- qcinjs_dat[, lapply(.SD, function(x) {sd(x, na.rm = T)/mean(x, na.rm = T)}), 
                        .SDcols = sph_list] %>% as.data.frame()
rownames(cv_qcinjs) <- "cv_qcinjs"
cv_qcinjs <- t(cv_qcinjs) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(cv_qcinjs, "rowname", "metabolite")
cv_qcinjs[, summary(cv_qcinjs)]

qcexts_dat[, (sph_list) := lapply(.SD, as.numeric), .SDcols = sph_list]
cv_qcexts <- qcexts_dat[, lapply(.SD, function(x) {sd(x, na.rm = T)/mean(x, na.rm = T)}), 
                        .SDcols = sph_list] %>% as.data.frame()
rownames(cv_qcexts) <- "cv_qcexts"
cv_qcexts <- t(cv_qcexts) %>% as.data.frame() %>% rownames_to_column() %>% as.data.table()
setnames(cv_qcexts, "rowname", "metabolite")
cv_qcexts[, summary(cv_qcexts)]

cv_qc <- merge(cv_qcinjs, cv_qcexts, by = "metabolite")

ggplot(cv_qc, aes(x = cv_qcinjs, y = cv_qcexts)) + 
        geom_point(size = 4, alpha = 0.5) + 
        labs(x = "CV of injection QC samples", y = "CV of extraction QC samples") + 
        xlim(0, 0.4) + ylim(0, 0.3) + 
        geom_abline(intercept = 0, slope = 1, linetype = 2, color = "darkgrey") + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "cv_qc_samples.png"), width = 16, height = 9)

ggplot(cv_qc, aes(x = cv_qcinjs, y = cv_qcexts)) + 
        geom_point(size = 4, alpha = 0.5) + 
        geom_label_repel(aes(label = ifelse(cv_qcinjs > 0.15 | cv_qcexts > 0.12, metabolite, "")),
                  box.padding   = 0.25, 
                  point.padding = 0.5,
                  segment.color = 'grey50') + 
        labs(x = "CV of injection QC samples", y = "CV of extraction QC samples") + 
        xlim(0, 0.4) + ylim(0, 0.3) + 
        geom_abline(intercept = 0, slope = 1, linetype = 2, color = "darkgrey") + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "cv_qc_samples_labeled.png"), width = 16, height = 9)

mets_info <- merge(mets_info, cv_qc, by = "metabolite")

```

# 3. Histograms & Correlations

```{r 3.hist_corr}

ggpairs(samp_dat[, ..sphoth_list], aes(alpha = 0.1)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "sphoth_pairs.png"), width = 20, height = 12)

sm_list_1 <- sm_list[1:10]
ggpairs(samp_dat[, ..sm_list_1], aes(alpha = 0.1)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "sm_part1_pairs.png"), width = 20, height = 12)

sm_list_2 <- sm_list[11:length(sm_list)]
ggpairs(samp_dat[, ..sm_list_2], aes(alpha = 0.1)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "sm_part2_pairs.png"), width = 20, height = 12)

cer_list_1 <- cer_list[1:13]
ggpairs(samp_dat[, ..cer_list_1], aes(alpha = 0.1)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "cer_part1_pairs.png"), width = 20, height = 12)

cer_list_2 <- cer_list[14:length(cer_list)]
ggpairs(samp_dat[, ..cer_list_2], aes(alpha = 0.1)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "cer_part2_pairs.png"), width = 20, height = 12)

ceroth_list_1 <- ceroth_list[1:12]
ggpairs(samp_dat[, ..ceroth_list_1], aes(alpha = 0.1)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "ceroth_part1_pairs.png"), width = 20, height = 12)

ceroth_list_2 <- ceroth_list[13:length(ceroth_list)]
ggpairs(samp_dat[, ..ceroth_list_2], aes(alpha = 0.1)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 16), 
              axis.text = element_text(size = 14))
ggsave(here(fig_dir, "ceroth_part2_pairs.png"), width = 20, height = 12)

```

# 4. Generate matched strata

```{r 4.strata}

strata <- samp_dat[Group == "Asthmatic", .(SubjectID, Match_ID)]
strata[, stratum := seq_along(SubjectID)]

samp_dat <- merge(samp_dat, strata[, .(SubjectID, stratum)], by = "SubjectID", all.x = T)

strata[, SubjectID := NULL]
strata[, SubjectID := Match_ID]

samp_dat <- merge(samp_dat, strata[, .(SubjectID, stratum)], by = "SubjectID", all.x = T)
samp_dat[1:20, .(asthma_sev, stratum.x, stratum.y)]

samp_dat[is.na(stratum.x), stratum.x := stratum.y]
samp_dat[, stratum := stratum.x]
samp_dat[, .N, stratum][, .N, N]
samp_dat[, (c("stratum.x", "stratum.y")) := NULL]

mets_pheno <- merge(samp_dat, pheno, by = "SubjectID", all.x = T)
unmatched_stratum <- mets_pheno[is.na(age), unique(stratum)]
mets_pheno[stratum == unmatched_stratum, .(SubjectID, stratum, Asthma_Status, Age, Gender, Race, BMI, 
                                           age, gender, race, bmi)] 

```

# 5. Table 1
## 5.1. Including all subjects in metabolomic data

```{r 5.1.table1_all}

mets_pheno[, .N, Comments_sample]
mets_pheno[, redness := case_when(Comments_sample == "very red" ~ "very red", 
                                  Comments_sample %in% c("red", "red and milky") ~ "red", 
                                  Comments_sample %in% c("light red", "milky and pink") ~ "light red/pink", 
                                  TRUE ~ "normal")]
mets_pheno[, milkiness := case_when(Comments_sample %in% c("milky", "red and milky", "milky and pink") ~ "milky", 
                                    TRUE ~ "normal")]
mets_pheno[, .N, .(Comments_sample, redness, milkiness)][order(milkiness, redness)]

cnt_cols <- c("Age", "BMI", "advair_prescriptions", "bronchodialator_prescriptions", 
              "food_allergy_diagnoses", "glucocorticoids_diagnoses", 
              "ige_value", "ige_measures", "igg_value", "igg_measures", 
              "inhaled_steriod_prescriptions", "oral_steroid_prescriptions", 
              "umeclidinium_measures", "unspecified_allergy_diagnoses")
cat_cols <- c("redness", "milkiness", "Gender", "Race", "Nonsmoker", 
              "allergic_rhinitis", "bronchodialator_nebulizer", 
              "cortisone", "food_allergy", "glucocorticoids", "hydrocortisone", "leukottriene", 
              "other_antiasthma", "pulmonary_lung_function", "theophylline", "triamcinolone", 
              "umeclidinium", "unspecified_allergy")

table1(cnt_cols, cat_cols, "mets_pheno", "Group")

```

## 5.2. Excluding the stratum containing the unmatched subject

```{r 5.2.table1_no_unmatched}

table1(cnt_cols, cat_cols, "mets_pheno[stratum != unmatched_stratum]", "Group")

mets_pheno[stratum != unmatched_stratum, .N, .(Group, inhaled_steriod_type)]
mets_pheno[stratum != unmatched_stratum, .N, .(Group, oral_steroid_type)]

```

# 6. Save data

```{r 6.save_data}

save(mets_pheno, mets_info, sph_list, sphoth_list, sm_list, cer_list, ceroth_list, 
     file = here(dat_dir, "eda_updated.RData"))

fwrite(mets_info, file = here(dat_dir, "metabolite_info.csv"))

```

# 7. Session info

```{r 7.sess_info}

sessionInfo()

```
